requirements.txt
anthropic==0.19.1
streamlit==1.30.0
yfinance==0.2.35
tenacity==8.2.3
langchain-core==0.1.20
langchain-openai==0.0.5
langgraph==0.0.33
python-dotenv==1.0.0
llama-index==0.10.57
llama-index-llms-openrouter==0.1.6
llama-index-embeddings-huggingface==0.2.2
sentence-transformers==2.7.0
llama-index-retrievers-bm25==0.1.3
pdfplumber==0.10.0
tabula-py==2.8.2
jpype1==1.5.0
pystemmer==2.2.0.1
nest-asyncio==1.6.0
